Leverage built-in validation highlighting.
In order to use the built-in validation highlighting stuff we have to add an attribute that will populate the `validate` attribute.  It is a bit of a hack but it gets the job done and allows us to remove the custom highlighting code, instead leveraging the bootstrap view strategy stuff.

Along with this, the validation now maps to all of the items without having to store the `ValidationGroups` separately or iterate over them when running `validate`.  This provides a minor improvement in readability and leverages Aurelia a bit more.