Adds a swappable OAuth for oauth.io.
The server is still expecting JWT tokens, so a full switch is pretty invasive.  However, auth0-lock is giant (500k minified) so it may be worth switching fully to oauth.io.

Note that this also explicitly doesn't bundle the OAuth-OAuthIo service, which causes the library to also not be bundled.  It isn't currently used at runtime and it is a non-trivial sized package (but still much smaller than Auth0).