Fixes a bug with multiple copies of JQuery on a page.
Apparently SystemJS doesn't do a perfect job of stopping JQuery from loading globally.  The net result is that we end up with two copies of JQuery on the page and when we reference it (via the popovers) we have to be very clear about which one we want to reference (the one that has the plugin popover plugin loaded in bootstrap.js).